Use chatbot web demo
